feat(auth): revamp login/signup pages with Opal components and pages/ structure
- Replace hand-rolled card/inputs/heading with Card, Content, InputTypeIn, InputVertical
- Use SvgOnyxLogo + "Welcome to Agent Wiki" title on both pages
- Extract logic into src/pages/LoginPage.tsx and src/pages/SignupPage.tsx
- Slim app/login/page.tsx and app/signup/page.tsx to bare re-exports
